+++ ethereum/evmtool/src/test/java/org/hyperledger/besu/evmtool/CodeValidationSubCommandTest.java
@@ -24,24 +24,24 @@ import java.io.PrintStream;
 import org.junit.jupiter.api.Test;
 import picocli.CommandLine;
 
-public class CodeValidationSubCommandTest {
+class CodeValidationSubCommandTest {
 
-  static final String CODE_STOP_ONLY = "0xef0001 010004 020001-0001 030000 00 00000000 00";
-  static final String CODE_RETF_ONLY = "0xef0001 010004 020001-0001 030000 00 00000000 e4";
-  static final String CODE_BAD_MAGIC = "0xefffff 010004 020001-0001 030000 00 00000000 e4";
+  static final String CODE_STOP_ONLY = "0xef0001 010004 020001-0001 040000 00 00800000 00";
+  static final String CODE_RETURN_ONLY = "0xef0001 010004 020001-0003 040000 00 00800002 5f5ff3";
+  static final String CODE_BAD_MAGIC = "0xefffff 010004 020001-0001 040000 00 00800000 e4";
   static final String CODE_INTERIOR_COMMENTS =
       """
-                  0xef0001 010008 020002-000c-0002 030000 00
+                  0xef0001 010008 020002-0009-0002 040000 00
                   # 7 inputs 1 output,
-                  00000007-07010007
-                  59-59-59-59-59-59-59-e30001-50-e4
+                  00800004-04010004
+                  59-59-59-59-e30001-50-00
                   # No immediate data
-                  f1-e4""";
+                  f8-e4""";
   static final String CODE_MULTIPLE =
-      CODE_STOP_ONLY + "\n" + CODE_BAD_MAGIC + "\n" + CODE_RETF_ONLY + "\n";
+      CODE_STOP_ONLY + "\n" + CODE_BAD_MAGIC + "\n" + CODE_RETURN_ONLY + "\n";
 
   @Test
-  public void testSingleValidViaInput() {
+  void testSingleValidViaInput() {
     final ByteArrayOutputStream baos = new ByteArrayOutputStream();
     final ByteArrayInputStream bais = new ByteArrayInputStream(CODE_STOP_ONLY.getBytes(UTF_8));
     final CodeValidateSubCommand codeValidateSubCommand =
@@ -51,7 +51,7 @@ public class CodeValidationSubCommandTest {
   }
 
   @Test
-  public void testSingleInvalidViaInput() {
+  void testSingleInvalidViaInput() {
     final ByteArrayOutputStream baos = new ByteArrayOutputStream();
     final ByteArrayInputStream bais = new ByteArrayInputStream(CODE_BAD_MAGIC.getBytes(UTF_8));
     final CodeValidateSubCommand codeValidateSubCommand =
@@ -61,7 +61,7 @@ public class CodeValidationSubCommandTest {
   }
 
   @Test
-  public void testMultipleViaInput() {
+  void testMultipleViaInput() {
     final ByteArrayOutputStream baos = new ByteArrayOutputStream();
     final ByteArrayInputStream bais = new ByteArrayInputStream(CODE_MULTIPLE.getBytes(UTF_8));
     final CodeValidateSubCommand codeValidateSubCommand =
@@ -72,12 +72,12 @@ public class CodeValidationSubCommandTest {
             """
                 OK 00
                 err: layout - EOF header byte 1 incorrect
-                OK e4
+                OK 5f5ff3
                 """);
   }
 
   @Test
-  public void testSingleValidViaCli() {
+  void testSingleValidViaCli() {
     final ByteArrayOutputStream baos = new ByteArrayOutputStream();
     final ByteArrayInputStream bais = new ByteArrayInputStream(new byte[0]);
     final CodeValidateSubCommand codeValidateSubCommand =
@@ -89,7 +89,7 @@ public class CodeValidationSubCommandTest {
   }
 
   @Test
-  public void testSingleInvalidViaCli() {
+  void testSingleInvalidViaCli() {
     final ByteArrayOutputStream baos = new ByteArrayOutputStream();
     final ByteArrayInputStream bais = new ByteArrayInputStream(new byte[0]);
     final CodeValidateSubCommand codeValidateSubCommand =
@@ -101,37 +101,37 @@ public class CodeValidationSubCommandTest {
   }
 
   @Test
-  public void testMultipleViaCli() {
+  void testMultipleViaCli() {
     final ByteArrayOutputStream baos = new ByteArrayOutputStream();
     final ByteArrayInputStream bais = new ByteArrayInputStream(new byte[0]);
     final CodeValidateSubCommand codeValidateSubCommand =
         new CodeValidateSubCommand(bais, new PrintStream(baos));
     final CommandLine cmd = new CommandLine(codeValidateSubCommand);
-    cmd.parseArgs(CODE_STOP_ONLY, CODE_BAD_MAGIC, CODE_RETF_ONLY);
+    cmd.parseArgs(CODE_STOP_ONLY, CODE_BAD_MAGIC, CODE_RETURN_ONLY);
     codeValidateSubCommand.run();
     assertThat(baos.toString(UTF_8))
         .contains(
             """
                 OK 00
                 err: layout - EOF header byte 1 incorrect
-                OK e4
+                OK 5f5ff3
                 """);
   }
 
   @Test
-  public void testCliEclipsesInput() {
+  void testCliEclipsesInput() {
     final ByteArrayOutputStream baos = new ByteArrayOutputStream();
     final ByteArrayInputStream bais = new ByteArrayInputStream(CODE_STOP_ONLY.getBytes(UTF_8));
     final CodeValidateSubCommand codeValidateSubCommand =
         new CodeValidateSubCommand(bais, new PrintStream(baos));
     final CommandLine cmd = new CommandLine(codeValidateSubCommand);
-    cmd.parseArgs(CODE_RETF_ONLY);
+    cmd.parseArgs(CODE_RETURN_ONLY);
     codeValidateSubCommand.run();
-    assertThat(baos.toString(UTF_8)).contains("OK e4\n");
+    assertThat(baos.toString(UTF_8)).contains("OK 5f5ff3\n");
   }
 
   @Test
-  public void testInteriorCommentsSkipped() {
+  void testInteriorCommentsSkipped() {
     final ByteArrayOutputStream baos = new ByteArrayOutputStream();
     final ByteArrayInputStream bais = new ByteArrayInputStream(new byte[0]);
     final CodeValidateSubCommand codeValidateSubCommand =
@@ -139,11 +139,11 @@ public class CodeValidationSubCommandTest {
     final CommandLine cmd = new CommandLine(codeValidateSubCommand);
     cmd.parseArgs(CODE_INTERIOR_COMMENTS);
     codeValidateSubCommand.run();
-    assertThat(baos.toString(UTF_8)).contains("OK 59595959595959e3000150e4,f1e4\n");
+    assertThat(baos.toString(UTF_8)).contains("OK 59595959e300015000,f8e4\n");
   }
 
   @Test
-  public void testBlankLinesAndCommentsSkipped() {
+  void testBlankLinesAndCommentsSkipped() {
     final ByteArrayOutputStream baos = new ByteArrayOutputStream();
     final ByteArrayInputStream bais =
         new ByteArrayInputStream(("# comment\n\n#blank line\n\n" + CODE_MULTIPLE).getBytes(UTF_8));
@@ -155,7 +155,7 @@ public class CodeValidationSubCommandTest {
             """
                 OK 00
                 err: layout - EOF header byte 1 incorrect
-                OK e4
+                OK 5f5ff3
                 """);
   }
 }