+++ ethereum/referencetests/src/main/java/org/hyperledger/besu/ethereum/referencetests/ReferenceTestProtocolSchedules.java
@@ -86,7 +86,7 @@ public class ReferenceTestProtocolSchedules {
     builder.put(
         "CancunToPragueAtTime15k",
         createSchedule(genesisStub.clone().cancunTime(0).pragueTime(15000)));
-    builder.put("Prague", createSchedule(genesisStub.clone().pragueTime(0)));
+    builder.put("Prague", createSchedule(genesisStub.clone().pragueEOFTime(0)));
     builder.put("Future_EIPs", createSchedule(genesisStub.clone().futureEipsTime(0)));
     builder.put("Experimental_EIPs", createSchedule(genesisStub.clone().experimentalEipsTime(0)));
     return new ReferenceTestProtocolSchedules(builder.build());