+++ evm/src/test/java/org/hyperledger/besu/evm/internal/CodeCacheTest.java
@@ -32,7 +32,7 @@ class CodeCacheTest {
     final Bytes contractBytes =
         Bytes.fromHexString("0xDEAD" + op + "BEEF" + op + "B0B0" + op + "C0DE" + op + "FACE");
     final CodeScale scale = new CodeScale();
-    final Code contractCode = CodeFactory.createCode(contractBytes, 0, false);
+    final Code contractCode = CodeFactory.createCode(contractBytes, 0);
     final int weight = scale.weigh(contractCode.getCodeHash(), contractCode);
     assertThat(weight)
         .isEqualTo(contractCode.getCodeHash().size() + (contractBytes.size() * 9 + 7) / 8);